Make game collection use promises. Delete some README files. Add a game model.
Rename GameCollection to OwnedGameCollection to distinguish between different endpoints in the
bgg api. Give it a method called fetchWithRetries that returns a $ promise which resolves when the
collection finally gets data instead of a "still processing" response.

Also: delete some README placeholders. Add a GameModel that doesn't do anything yet. Add some
diagnostic behavior to the index which will be changed later.